+++ R/dictionaries.R
@@ -224,9 +224,11 @@ dictionary.default <- function(x, file = NULL, format = NULL,
         x <- select_dictionary_levels(x, levels)
     if (tolower)
         x <- lowercase_dictionary_values(x)
-    x <- merge_dictionary_values(x)
     if (tokenize)
         x <- tokenize_dictionary_values(x, separator)
+    if (separator != " ")
+        x <- replace_dictionary_values(x, separator, " ")
+    x <- merge_dictionary_values(x)
     build_dictionary2(x, separator = separator)
 }
 
@@ -249,10 +251,11 @@ dictionary.list <- function(x, file = NULL, format = NULL,
         x <- select_dictionary_levels(x, levels)
     if (tolower)
         x <- lowercase_dictionary_values(x)
-    x <- replace_dictionary_values(x, separator, " ")
-    x <- merge_dictionary_values(x)
     if (tokenize)
         x <- tokenize_dictionary_values(x, separator)
+    if (separator != " ")
+        x <- replace_dictionary_values(x, separator, " ")
+    x <- merge_dictionary_values(x)
     build_dictionary2(x, separator = separator)
 }
 
@@ -273,7 +276,7 @@ dictionary.dictionary2 <- function(x, file = NULL, format = NULL,
 #' @param ... passed to [quanteda::tokens()].
 #' @keywords dictionary development internal
 #' @export
-dictionary_tokenize <- function(dictionary, ...) {
+tokenize_dictionary <- function(dictionary, ...) {
 
     if (!is.dictionary(dictionary))
         stop("Dictionary object must be provided")
@@ -542,27 +545,27 @@ setMethod("c",
 
 #' Internal function for special handling of multi-word dictionary values
 #' @param dict a flatten dictionary
-#' @param concatenator_dictionary concatenator from a dictionary object
-#' @param concatenator_tokens concatenator from a tokens object
+#' @param separator separator from a dictionary object
+#' @param concatenator concatenator from a tokens object
 #' @keywords internal
 #' @importFrom stringi stri_detect_fixed stri_split_fixed stri_replace_all_fixed
-split_values <- function(dict, concatenator_dictionary, concatenator_tokens) {
+split_values <- function(dict, separator, concatenator) {
 
     key <- rep(names(dict), lengths(dict))
     value <- unlist(dict, use.names = FALSE)
-    is_multi <- stri_detect_fixed(value, concatenator_dictionary)
+    is_multi <- stri_detect_fixed(value, separator)
     if (any(is_multi)) {
         result <- vector("list", length(value) + sum(is_multi))
         l <- !seq_along(result) %in% (seq_along(value) + cumsum(is_multi))
-        result[l] <- stri_split_fixed(value[is_multi], concatenator_dictionary)
+        result[l] <- stri_split_fixed(value[is_multi], separator)
         result[!l] <- as.list(stri_replace_all_fixed(value,
-                                                     concatenator_dictionary,
-                                                     concatenator_tokens))
+                                                     separator,
+                                                     concatenator))
         names(result) <- key[rep(seq_along(value), 1 + is_multi)]
     } else {
         result <- as.list(stri_replace_all_fixed(value,
-                                                 concatenator_dictionary,
-                                                 concatenator_tokens))
+                                                 separator,
+                                                 concatenator))
         names(result) <- key
     }
     return(result)