+++ tests/testthat/test-dictionaries.R
@@ -644,7 +644,7 @@ test_that("tokenize is working", {
     expect_equivalent(
         dict4,
         list(ASIA = list("IN" = "印度", 
-                         "ID" = "印度_尼西亚"))
+                         "ID" = "印度 尼西亚"))
     )
     
     # with wildcard
@@ -710,7 +710,7 @@ test_that("tokenize is working", {
 test_that("levels is working", {
     
     dict <- dictionary(file = "../data/dictionaries/newsmap2.yml", 
-                       tolower = FALSE, levels = 1:2)
+                       tolower = FALSE, levels = 1:2, separator = "_")
 
     # dictionary
     
@@ -788,7 +788,8 @@ test_that("as.yaml is commutative", {
         list("US" = list("Washington DC",
                          "MA" = "Boston",
                          "CA" = "Sacramento"),
-             "JP" = list("Tokyo")), tolower = FALSE)
+             "JP" = list("Tokyo")), 
+        tolower = FALSE)
     
     # write and read
     f <- paste0(tempfile(), ".yml")
@@ -797,5 +798,36 @@ test_that("as.yaml is commutative", {
     expect_identical(
         dict, 
         dictionary(file = f, tolower = FALSE)
+    )
+    
+})
+
+
+test_that("separator is working", {
+    
+    dict1 <- dictionary(
+        list("US" = list("Washington_DC",
+                         "MA" = "Boston",
+                         "CA" = "Sacramento"),
+             "JP" = list("Tokyo")), 
+        tolower = FALSE, separator = "_")
+    
+    expect_equal(dict1@meta$object$separator, "_")
+    expect_true(
+        any(stri_detect_fixed(unlist(dict1), " "))
+    )
+    expect_false(
+        any(stri_detect_fixed(unlist(dict1), "_"))
+    )
+    
+    dict2 <- dictionary(file = "../data/dictionaries/newsmap2.yml", 
+                        tolower = FALSE, separator = "_")
+    
+    expect_equal(dict2@meta$object$separator, "_")
+    expect_true(
+        any(stri_detect_fixed(unlist(dict2), " "))
+    )
+    expect_false(
+        any(stri_detect_fixed(unlist(dict2), "_"))
     )
 })